Move charset generation into package
